+++ pkg/alertmanager/amcfg.go
@@ -1210,6 +1210,10 @@ func (cb *configBuilder) convertTelegramConfig(ctx context.Context, in monitorin
 	}
 	out.HTTPConfig = httpConfig
 
+	if in.MessageThreadID != nil {
+		out.MessageThreadID = int(*in.MessageThreadID)
+	}
+
 	if in.BotToken != nil {
 		botToken, err := cb.store.GetSecretKey(ctx, crKey.Namespace, *in.BotToken)
 		if err != nil {
@@ -2256,6 +2260,12 @@ func (tc *telegramConfig) sanitize(amVersion semver.Version, logger *slog.Logger
 		msg := "'bot_token' and 'bot_token_file' are mutually exclusive for telegram receiver config - 'bot_token' has taken precedence"
 		logger.Warn(msg)
 		tc.BotTokenFile = ""
+	}
+
+	if tc.MessageThreadID != 0 && lessThanV0_26 {
+		msg := "'message_thread_id' supported in Alertmanager >= 0.26.0 only - dropping field from provided config"
+		logger.Warn(msg, "current_version", amVersion.String())
+		tc.MessageThreadID = 0
 	}
 
 	return tc.HTTPConfig.sanitize(amVersion, logger)