+++ pkg/client/applyconfiguration/monitoring/v1/commonprometheusfields.go
@@ -114,6 +114,7 @@ type CommonPrometheusFieldsApplyConfiguration struct {
 	ServiceDiscoveryRole                 *monitoringv1.ServiceDiscoveryRole                      `json:"serviceDiscoveryRole,omitempty"`
 	TSDB                                 *TSDBSpecApplyConfiguration                             `json:"tsdb,omitempty"`
 	ScrapeFailureLogFile                 *string                                                 `json:"scrapeFailureLogFile,omitempty"`
+	Runtime                              *RuntimeConfigApplyConfiguration                        `json:"runtime,omitempty"`
 }
 
 // CommonPrometheusFieldsApplyConfiguration constructs a declarative configuration of the CommonPrometheusFields type for use with
@@ -889,5 +890,13 @@ func (b *CommonPrometheusFieldsApplyConfiguration) WithTSDB(value *TSDBSpecApply
 // If called multiple times, the ScrapeFailureLogFile field is set to the value of the last call.
 func (b *CommonPrometheusFieldsApplyConfiguration) WithScrapeFailureLogFile(value string) *CommonPrometheusFieldsApplyConfiguration {
 	b.ScrapeFailureLogFile = &value
+	return b
+}
+
+// WithRuntime sets the Runtime field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the Runtime field is set to the value of the last call.
+func (b *CommonPrometheusFieldsApplyConfiguration) WithRuntime(value *RuntimeConfigApplyConfiguration) *CommonPrometheusFieldsApplyConfiguration {
+	b.Runtime = value
 	return b
 }